Refactor files into python modules.
